requirements.txt
pybit==5.6.2
pandas>=2.2.0
numpy>=1.26.0
ta==0.11.0
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0

# AWS & RabbitMQ
boto3==1.34.*
pika==1.3.*
